refactor: rename urban, fix relative imports
* Rename urban to __main__.py

* Rename src to urban

* Get main function working

Issue: #55
Tag: 2.0.0

Authored-by: Joshua Rose <joshuarose099@gmail.com>
