--- src/urban.py
@@ -1,13 +1,0 @@
-"""
-Centralized urban file that is calls the API.
-"""
-
-from urban_api import apply_word_to_url, send_exists_request, send_phrase_request
-
-
-def main():
-    pass
-
-
-if __name__ == "__main__":
-    main()
